Rewriting write_example
This new version accepts more command line parameters to alter the
different execution options involved in writing to the casa table. The
new options include the cell size, the number of dimensions to use for
the cell (only 1 and 2 are supported at the moment), the data type (with
float, double, complex and dcomplex being supported), and the number of
rows to write. The default values for all of these options are those
previously hardcoded in the code.

The code has also been rearranged a bit to make exception handling a bit
easier, use a random value to fill the data arrays, and have easier
memory handling.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>